# backend/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "AKMC ADmin API",
    "description": "API documentation for AKMC admin apis"
  },
  "host": "localhost:5000",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "Authorization": {
      "type": "apiKey",
      "name": "Authorization",
      "description": "Value: Bearer ",
      "in": "header",
      "scheme": "bearer"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/LoginModel"
            }
          }
        ],
        "responses": {}
      }
    },
    "/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/RegisterModel"
            }
          }
        ],
        "responses": {
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/auth/verify": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/VerifyEmailModel"
            }
          }
        ],
        "responses": {}
      }
    },
    "/auth/forgotPassword": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ForgotPassWordModel"
            }
          }
        ],
        "responses": {}
      }
    },
    "/auth/resetPassword": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ResetPasswordModel"
            }
          }
        ],
        "responses": {}
      }
    },
    "/auth/changePassword": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ChangePasswordModel"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
...
  },
  "definitions": {
    "LoginModel": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "example": "admin@markscodingspot.com"
        },
        "password": {
          "type": "string",
          "example": "Password123#"
        }
      },
      "required": [
        "email",
        "password"
      ]
    },
    "RegisterModel": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "John Snow"
        },
        "email": {
          "type": "string",
          "example": "mail@mail.com"
        },
        "password": {
          "type": "string",
          "example": "Password123#"
        }
      },
      "required": [
        "name",
        "email",
        "password"
      ]
    },
...
    "VerifyEmailModel": {
      "type": "object",
      "properties": {
        "code": {
          "type": "number",
          "example": 333333
        }
      },
      "required": [
        "code"
      ]
    },
    "ChangePasswordModel": {
      "type": "object",
      "properties": {
        "oldPassword": {
          "type": "string",
          "example": "Password123#"
        },
        "newPassword": {
          "type": "string",
          "example": "Password789#"
        }
      },
      "required": [
        "oldPassword",
        "newPassword"
      ]
    },
    "ForgotPassWordModel": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "example": "coommark@gmail.com"
        }
      },
      "required": [
        "email"
      ]
    },
    "ResetPasswordModel": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "example": "coommark@gmail.com"
        },
        "code": {
          "type": "number",
          "example": 999999
        },
        "newPassword": {
          "type": "string",
          "example": "Password789#"
        }
      },
      "required": [
        "email",
        "code",
        "newPassword"
      ]
    }
  }
}